+++ actions/tests/actions_tests/test_actions_response.py
@@ -7,8 +7,8 @@ def test_actions_response_list(datadir, data_regression):
         ["list", "--skip-lint"], returncode=0, cwd=str(datadir)
     )
     found = json.loads(result.stdout)
-    assert len(found) == 4
     # print(json.dumps(found, indent=4))
+    assert len(found) == 5
     data = {}
     for f in found:
         data[f["name"]] = {
@@ -55,11 +55,25 @@ def test_actions_raise_other_error_action(datadir, data_regression):
     assert "FAIL" in result.stdout.decode("utf-8")
 
 
-def test_actions_rase_action_error_action(datadir, data_regression):
+def test_actions_raise_error_bad_schema(datadir, data_regression):
     from devutils.fixtures import sema4ai_actions_run
 
     result = sema4ai_actions_run(
-        ["run", "-a=rase_action_error_action", "--print-result"],
+        ["run", "-a=bad_schema", "--print-result"],
+        returncode=0,
+        cwd=str(datadir),
+    )
+    assert (
+        "Although the action: 'bad_schema' ran properly, it returned a value of type <class 'int'> whereas the expected return type is <class 'sema4ai.actions._response.Response[str]'>."
+        in result.stdout.decode("utf-8")
+    )
+
+
+def test_actions_raise_action_error_action(datadir, data_regression):
+    from devutils.fixtures import sema4ai_actions_run
+
+    result = sema4ai_actions_run(
+        ["run", "-a=raise_action_error_action", "--print-result"],
         returncode=1,
         cwd=str(datadir),
     )